requirements.txt
# OATD Thesis Crawler Dependencies
# Core - only lightweight libs needed since we use opencli + Chrome
pyyaml>=6.0.1
loguru>=0.7.0
rich>=13.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
